requirements.txt
antlr4-python3-runtime
networkx
matplotlib
python-telegram-bot